refactor: votes
