+++ internal/domain/postmethods.go
@@ -6,7 +6,6 @@ type PostUsecase interface {
 	GetPostByID(id int64) (*Post, error)
 	GetPostsByCategory(category string) ([]*PostDTO, error)
 	GetAllPosts() ([]*PostDTO, error)
-	VotePost(postID, userID int64, vote int) error
 	GetVotedPostsByUserID(userID int64) ([]*PostDTO, error)
 	GetVotesCountByPostID(postID int64) (*Vote, error)
 }
@@ -17,7 +16,6 @@ type PostRepo interface {
 	GetPostByID(id int64) (*Post, error)
 	GetPostsByCategory(category string) ([]*PostDTO, error)
 	GetAllPosts() ([]*PostDTO, error)
-	VotePost(postID, userID int64, vote int) error
 	GetVotedPostsByUserID(userID int64) ([]*PostDTO, error)
 	GetVotesCountByPostID(postID int64) (*Vote, error)
 }
