+++ internal/domain/votemethods.go
@@ -1,0 +1,11 @@
+package domain
+
+type VoteUsecase interface {
+	VotePost(postID, userID int64, vote int) error
+	VoteComment(commentID, userID int64, vote int) error
+}
+
+type VoteRepo interface {
+	VotePost(postID, userID int64, vote int) error
+	VoteComment(commentID, userID int64, vote int) error
+}
